-Use existing log helper -exclude messaging, constants from testing -Use existing app properties
